Restructure files and debounce on change quantity
