+++ package.json
@@ -16,6 +16,8 @@
     "apollo-link-context": "^1.0.8",
     "babel-loader": "^8.0.2",
     "graphql-tag": "^2.9.0",
+    "lodash": "^4.17.11",
+    "lodash.debounce": "^4.0.8",
     "register-service-worker": "^1.0.0",
     "vue": "^2.5.17",
     "vue-apollo": "^3.0.0-beta.25",
